// types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      project_techs: {
        Row: {
          project_id: string
          tech_id: string
        }
        Insert: {
          project_id: string
          tech_id: string
        }
        Update: {
          project_id?: string
          tech_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "project_techs_project_id_fkey"
            columns: ["project_id"]
            isOneToOne: false
            referencedRelation: "projects"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "project_techs_tech_id_fkey"
            columns: ["tech_id"]
            isOneToOne: false
            referencedRelation: "tech_stack"
            referencedColumns: ["id"]
          },
        ]
      }
      projects: {
        Row: {
          content_en: string | null
          content_es: string | null
          created_at: string
          description_en: string
          description_es: string
          display_order: number
          id: string
          image_url: string
          is_featured: boolean | null
          live_url: string | null
          repo_url: string | null
          title_en: string
          title_es: string
        }
        Insert: {
          content_en?: string | null
          content_es?: string | null
          created_at?: string
          description_en: string
          description_es: string
          display_order?: number
          id?: string
          image_url: string
          is_featured?: boolean | null
          live_url?: string | null
          repo_url?: string | null
          title_en: string
          title_es: string
        }
        Update: {
          content_en?: string | null
          content_es?: string | null
          created_at?: string
          description_en?: string
          description_es?: string
          display_order?: number
          id?: string
          image_url?: string
          is_featured?: boolean | null
          live_url?: string | null
          repo_url?: string | null
          title_en?: string
          title_es?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}